split disposable_js into modal_js and static_plots
